Crazy crazy merger fix.

+++ src/ui/containers/HomeStatusContainer.java
@@ -8,10 +8,10 @@ import java.time.DayOfWeek;
  * Created by fegwin on 13/10/2016.
  */
 public class HomeStatusContainer extends StatusContainerBase {
-    public int currentNetConsumption;
+    public double currentNetConsumption;
     public Proposal currentEnergyContract;
 
-    public HomeStatusContainer(int currentNetConsumption, int hourOfDay, DayOfWeek dayOfWeek, Proposal contract) {
+    public HomeStatusContainer(double currentNetConsumption, int hourOfDay, DayOfWeek dayOfWeek, Proposal contract) {
         super(hourOfDay, dayOfWeek);
 
         this.currentNetConsumption = currentNetConsumption;
